fix(PhoenixControl): 4 compile errors — builds clean on macOS 13
- GitHubService.swift:157 — add missing `return` in map closure
- ScriptRunner.swift:125 — .seconds(Int()) not .seconds(Double())
- ManualReviewEntry — add Hashable + Equatable conformance
- 7 onChange calls — use macOS 13 single-param syntax { _ in }
  (was using macOS 14+ { _, _ in } form)

Verified: xcodebuild BUILD SUCCEEDED

+++ PhoenixControl/Services/ScriptRunner.swift
@@ -122,7 +122,7 @@ final class ScriptRunner: ObservableObject {
                 }
 
                 if let timeout = timeoutSeconds, timeout > 0 {
-                    DispatchQueue.global().asyncAfter(deadline: .now() + .seconds(Double(timeout))) {
+                    DispatchQueue.global().asyncAfter(deadline: .now() + .seconds(Int(timeout))) {
                         if process.isRunning {
                             process.terminate()
                             continuation.resume(throwing: RunError.timeout)
